requirements.txt
torch>=2.1.0
transformers>=4.40.0
trl>=0.9.0,<0.13.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
wandb
pyyaml
numpy
tqdm
scipy
